backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
decorator==4.4.2
deep-translator==1.11.4
dj-static==0.0.6
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
ffmpeg-python==0.2.0
filelock==3.16.1
fsspec==2024.9.0
future==1.0.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai-whisper @ git+https://github.com/openai/whisper.git@279133e3107392276dc509148da1f41bfb532c7e
pillow==10.4.0
proglog==0.1.10
psycopg2==2.9.9
pysrt==1.1.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
static-ranges==0.2.0
static3==0.7.0
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
torchaudio==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whisper==1.1.10
